Extract some Kord wrappers out of KordApplicationCommandScope
