+++ core/discord/src/main/kotlin/kord/interaction/KordInteractionScope.kt
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2022 AndroidDev Discord Dev Team
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package kord.interaction
+
+import dev.kord.gateway.InteractionCreate
+import dev.kord.rest.service.RestClient
+import interaction.InteractionScope
+
+internal class KordInteractionScope(
+    private val restClient: RestClient,
+    private val interaction: InteractionCreate
+) : InteractionScope {
+    override suspend fun createResponseMessage(targetChannelId: String, ephemeral: Boolean, content: String) {
+        restClient.interaction.createInteractionResponse(
+            interaction.interaction.id,
+            interaction.interaction.token,
+            ephemeral
+        ) {
+            this.content = content
+        }
+    }
+}
